# Conflicts:
#	models/_settings.json
#	models/wan/wan_handler.py
#	shared/utils/utils.py
#	wgp.py

+++ models/wan/scail/model_scail.py
@@ -32,7 +32,7 @@ def build_scail_pose_tokens(model, pose_latents: torch.Tensor, target_dtype: Opt
         dtype=pose_latents.dtype,
     )
     pose_input = torch.cat([pose_latents, pose_mask], dim=1)
-    pose_embed = model.pose_patch_embedding(pose_input)  # (B, dim, T', H', W')
+    pose_embed = model.pose_patch_embedding(pose_input.to(model.pose_patch_embedding.weight.dtype))  # (B, dim, T', H', W')
     return pose_embed.flatten(2).transpose(1, 2)
